+++ sentry-ruby/lib/sentry/net/http.rb
@@ -30,13 +30,13 @@ module Sentry
         return super if from_sentry_sdk?
 
         Sentry.with_child_span(op: OP_NAME, start_timestamp: Sentry.utc_now.to_f) do |sentry_span|
-          set_sentry_trace_header(req, sentry_span)
+          request_info = extract_request_info(req)
+          set_sentry_trace_header(req, sentry_span, request_info)
 
           super.tap do |res|
-            record_sentry_breadcrumb(req, res)
+            record_sentry_breadcrumb(request_info, res)
 
             if sentry_span
-              request_info = extract_request_info(req)
               sentry_span.set_description("#{request_info[:method]} #{request_info[:url]}")
               sentry_span.set_data('url', request_info[:url])
               sentry_span.set_data('http.method', request_info[:method])
@@ -49,10 +49,11 @@ module Sentry
 
       private
 
-      def set_sentry_trace_header(req, sentry_span)
+      def set_sentry_trace_header(req, sentry_span, request_info)
         return unless sentry_span
 
         client = Sentry.get_current_client
+        return unless propagate_trace?(request_info[:url], client.configuration.trace_propagation_targets)
 
         trace = client.generate_sentry_trace(sentry_span)
         req[SENTRY_TRACE_HEADER_NAME] = trace if trace
@@ -61,10 +62,8 @@ module Sentry
         req[BAGGAGE_HEADER_NAME] = baggage if baggage && !baggage.empty?
       end
 
-      def record_sentry_breadcrumb(req, res)
+      def record_sentry_breadcrumb(request_info, res)
         return unless Sentry.initialized? && Sentry.configuration.breadcrumbs_logger.include?(:http_logger)
-
-        request_info = extract_request_info(req)
 
         crumb = Sentry::Breadcrumb.new(
           level: :info,
@@ -95,6 +94,10 @@ module Sentry
         end
 
         result
+      end
+
+      def propagate_trace?(url, trace_propagation_targets)
+        url && trace_propagation_targets.any? { |target| url.match?(target) }
       end
     end
   end